+++ tests/test_day_boundary.py
