+++ tests/test_energy_aggregator.py
